requirements.txt
fastapi
uvicorn
python-multipart
pypdf
groq
python-dotenv
motor
beanie
requests
openpyxl
reportlab
